Add FileParser to make multi-file clearer
